requirements-file.txt
langchain>=0.0.267
openai>=1.3.0
chromadb>=0.4.13
tiktoken>=0.5.1
pypdf>=3.15.1
python-dotenv>=1.0.0
